Renamed all config vars, split roles into separate file
Role config vars in roles/*/vars/main.yml are now `role_name__var`,
and can be overridden by setting them in config.yml.

Roles are now stored in a separate file from the rest of the config to
make it easier to auto-generate a list of roles to include based on the
contents of the roles/ dir, though this has not yet been implemented.