Added roles back into main.yml
I split the roles list into a separate file, but apparently this
prevents the roles from seeing (at least) the handlers defined in
main.yml

It would still be good to have a way to dynamically generate the list of
roles to be used, but this isn't it. Instead, we should probably start
using a Makefile or similar to generate main.yml from a template.