Change rsync syntax
The rsync:// syntax doesn't work with the geekdome.net rsyncd setup for
some reason, but server::module works with both geekdome and
worldchanging.org, so switching to that.

I also removed the trailing '/' from the source path. This way you have
a choice of whether to sync only the contents of the directory (with
'/'), or the directory its self as a subdir of the destination (without
'/').