Make wikipedia role use rachel module
This retires the old apache-based wikipedia role and replaces it with
(for now) a very simple role that just deploys the WfS Rachel module.

TODO: Make wikipedia.org URLs redirect to the appropriate WfS page (and
ideally try the actual Wikipdedia if that doesn't work and net access is
available).