Replace bind with dnsmasq, create generic dns_server role
dns_server acts as a generic API of sorts: dependent roles can call it, give it information about e.g. aliases they want added in an abstract way, which it passes to whatever service it's configured to use. Only dnsmasq is currently supported, but this will let us swap to something else with minimal impact if we ever need to.